scxtop: Fix perf events process filtering
Fix the api for PerfEvent to take a i32 to follow the kernel API to
easily allow for passing in -1 (all CPUs). Update the MCP server to fix
perf event filtering for processes. Add tests for related changes.

+++ tools/scxtop/src/profiling_events/mod.rs
@@ -38,7 +38,7 @@ impl ProfilingEvent {
         match self {
             ProfilingEvent::Perf(p) => {
                 let mut p = p.clone();
-                p.cpu = cpu;
+                p.cpu = cpu as i32;
                 p.attach(process)?;
                 Ok(ProfilingEvent::Perf(p))
             }
